### 2023-09-13 #5845 @oliviaguest
Hey, thanks for submitting to JOSS, @fredshone. I am going to try and find an editor for you, but can you also check the missing refs while I do that, please? Thanks!
