+++ resources/paper/paper.bib
@@ -1,9 +1,11 @@
 @book{MATSim:2016,
   	Author = {{Horni}, A. and {Nagel}, K. and {Axhausen}, K.W. (eds.)},
-  	Booktitle = {2016 The Multi-Agent Transport Simulation MATSim. DOI: http://dx.doi.org/10.5334/baw. License: CC-BY 4.0},
+  	Booktitle = {2016 The Multi-Agent Transport Simulation MATSim. License: CC-BY 4.0},
   	Publisher = {London: Ubiquity Press.},
   	Title = {{The Multi-Agent Transport Simulation MATSim}},
-  	Year = 2016
+  	Year = 2016,
+    doi = {10.5334/baw},
+    url = {http://dx.doi.org/10.5334/baw}
 }
 
 @book{TRB-primer,
@@ -11,6 +13,7 @@
   	Title = {Activity-Based Travel Demand Models: A Primer},
   	Publisher = {Washington, DC: The National Academies Press},
   	Year = 2014,
+    doi = {10.17226/22357},
     url = {https://doi.org/10.17226/22357}
 }
 
@@ -19,6 +22,7 @@
   	Title = {Activity-based models of travel demand: promises, progress and prospects},
   	Publisher = {International Journal of Urban Sciences},
   	Year = 2014,
+    doi = {10.1080/12265934.2013.835118},
     url = {https://doi.org/10.1080/12265934.2013.835118}
 }
 
