### 2023-09-13 b1ca93e
discretionary location choice
