+++ pam/planner/choice_location.py
@@ -4,12 +4,12 @@ import logging
 from collections.abc import Callable
 from copy import deepcopy
 from dataclasses import dataclass
-from typing import Literal, NamedTuple, Optional, Union
+from typing import Literal, NamedTuple, Optional, Union, List
 
 import numpy as np
 import pandas as pd
 
-from pam.activity import Activity
+from pam.activity import Activity, Leg, Plan
 from pam.core import Population
 from pam.operations.cropping import link_population
 from pam.planner.od import OD
@@ -17,6 +17,10 @@ from pam.planner.utils_planner import (
     apply_mode_to_home_chain,
     calculate_mnl_probabilities,
     sample_weighted,
+    get_first_leg_time_ratio,
+    get_act_names,
+    convert_single_anchor_roundtrip,
+    get_trip_chains_either_anchor
 )
 from pam.planner.zones import Zones
 
@@ -253,3 +257,222 @@ class ChoiceMNL(ChoiceModel):
         self.configure(
             func_probabilities=calculate_mnl_probabilities, func_sampling=sample_weighted
         )
+
+
+class DiscretionaryTrips:
+    """
+    Solve discretionary trip location choice of a PAM plan
+    """
+
+    def __init__(self, plan: Plan, od: OD) -> None:
+        self._plan = plan
+        self._od = od
+
+    def update_plan(self):
+        trip_chains = get_trip_chains_either_anchor(self._plan)
+        for trip_chain in trip_chains:
+            convert_single_anchor_roundtrip(trip_chain) # if only one achor, convert to round-trip
+            act_names = get_act_names(trip_chain)
+            if len(act_names) > 2:
+                if act_names[0]!=act_names[-1]:
+                    DiscretionaryTripOD(
+                        trip_chain=trip_chain,
+                        od=self._od
+                    ).update_plan()
+                else:                    
+                    DiscretionaryTripRound(
+                        trip_chain=trip_chain,
+                        od=self._od
+                    ).update_plan()
+
+
+class DiscretionaryTrip:
+    """
+    Location choice for discretionary trips
+        in a trip chain
+
+    Cases:
+        a) O->discretionary->O (DiscretionaryTripRound)
+        b) O->discretionary->D (DiscretionaryTripOD)
+        c) O->discretionary->discretionary->O
+        d) O->discretionary->discretionary->D
+
+    Chains with multiple discretionary trips are solved 
+        recursively, updating the first location each time,
+        and then keeping it fixed as we solve downstream.       
+    """
+
+    def __init__(
+        self,
+        trip_chain: List[Union[Activity, Leg]],
+        od: OD
+    ) -> None:
+        self._trip_chain = trip_chain
+        self._od = od
+        self.act_names = get_act_names(trip_chain)
+
+        # anchor points
+        self.anchor_zone_start = trip_chain[0].location.area
+        self.anchor_zone_end = trip_chain[-1].location.area
+        self.trmode = trip_chain[1].mode
+
+        # self.anchor_zone_start = self._od.encode(self.anchor_area_start)
+        # self.anchor_zone_end = self._od.encode(self.anchor_area_end)
+
+    def choose_destination(self):
+        # implemented in the DiscretionaryTripRound and
+        # DiscretionaryTripOD subclasses
+        raise NotImplementedError
+
+    def update_plan(self):
+        """
+        Update the PAM activity locations of the first activity activity 
+            in the trip chain,
+            and continue downstream until the entire chain is solved.
+        """
+        if len(self.act_names) > 2:
+            assert isinstance(self._trip_chain[0], Activity)
+            assert isinstance(self._trip_chain[1], Leg)
+            assert isinstance(self._trip_chain[2], Activity)
+            assert isinstance(self._trip_chain[3], Leg)
+
+            # update locations
[…]
