+++ pam/variables.py
@@ -20,4 +20,4 @@ EXPECTED_EUCLIDEAN_SPEEDS = {
 TRANSIT_MODES = ["bus", "rail", "pt", "subway"]  # modes for which the maximum walk distance applies
 SMALL_VALUE = 0.000001
 
-LONG_TERM_ACTIVITIES = ["work", "school", "education"]
+LONG_TERM_ACTIVITIES = ["work", "school", "education", "home"]

[…]
