+++ pam/planner/choice_location.py
@@ -67,7 +67,8 @@ class SelectionSet:
 
     def sample(self) -> list:
         """Sample from a set of alternative options."""
-        sampled = np.apply_along_axis(func1d=self.func_sampling, axis=1, arr=self.probabilities)
+        sampled = np.apply_along_axis(
+            func1d=self.func_sampling, axis=1, arr=self.probabilities)
         sampled_labels = [self.choice_set.choice_labels[x] for x in sampled]
         self._selections = sampled_labels
         return sampled_labels
@@ -128,7 +129,8 @@ class ChoiceModel:
         link_population(self.population)
         self.od = od
         self.zones = self.parse_zone_data(zones)
-        self.zones.data = self.zones.data.loc[list(od.labels.destination_zones)]
+        self.zones.data = self.zones.data.loc[list(
+            od.labels.destination_zones)]
         self.configuration = ChoiceConfiguration()
         self._selections = None
 
@@ -212,7 +214,8 @@ class ChoiceModel:
 
         idxs = []
         u_choices = []
-        choice_labels = list(itertools.product(od.labels.destination_zones, od.labels.mode))
+        choice_labels = list(itertools.product(
+            od.labels.destination_zones, od.labels.mode))
         choice_labels = [ChoiceLabel(*x) for x in choice_labels]
 
         # iterate across activities
@@ -269,17 +272,23 @@ class DiscretionaryTrips:
         self._od = od
 
     def update_plan(self):
+        """Update the locations (in-place) of each non-mandatory activity location in the plan.
+
+        Returns:
+            None
+        """
         trip_chains = get_trip_chains_either_anchor(self._plan)
         for trip_chain in trip_chains:
-            convert_single_anchor_roundtrip(trip_chain) # if only one achor, convert to round-trip
+            # if only one achor, convert to round-trip
+            convert_single_anchor_roundtrip(trip_chain)
             act_names = get_act_names(trip_chain)
             if len(act_names) > 2:
-                if act_names[0]!=act_names[-1]:
+                if act_names[0] != act_names[-1]:
                     DiscretionaryTripOD(
                         trip_chain=trip_chain,
                         od=self._od
                     ).update_plan()
-                else:                    
+                else:
                     DiscretionaryTripRound(
                         trip_chain=trip_chain,
                         od=self._od
@@ -329,6 +338,10 @@ class DiscretionaryTrip:
         Update the PAM activity locations of the first activity activity 
             in the trip chain,
             and continue downstream until the entire chain is solved.
+
+        Returns:
+            None
+
         """
         if len(self.act_names) > 2:
             assert isinstance(self._trip_chain[0], Activity)
@@ -370,11 +383,14 @@ class DiscretionaryTripRound(DiscretionaryTrip):
     """
 
     def choose_destination(self) -> str:
-        """
-        Choose a destination for the discretionary activity
+        """Selects a destination for the discretionary activity
+
+        Returns:
+            str: Selected destination zone name.
         """
         assert isinstance(self._trip_chain[1], Leg)
-        destination_p = self._od['od_probs', self.anchor_zone_start, :, self.trmode]
+        destination_p = self._od['od_probs',
+                                 self.anchor_zone_start, :, self.trmode]
         destination_p = destination_p / destination_p.sum()
         zone = sample_weighted(destination_p)
         area = self._od.labels.destination_zones[zone]
@@ -415,9 +431,14 @@ class DiscretionaryTripOD(DiscretionaryTrip):
     def leg_ratios(self) -> np.array:
         """
         Get the impedance ratio between the fist leg candidate locations
-            and the corresponding total distance (anchor->discretionary->anchor)
+            and the corresponding total distance (anchor->discretionary->anchor).
+
+        Returns:
[…]
