+++ pam/planner/choice_location.py
@@ -4,7 +4,7 @@ import logging
 from collections.abc import Callable
 from copy import deepcopy
 from dataclasses import dataclass
-from typing import Literal, NamedTuple, Optional, Union, List
+from typing import List, Literal, NamedTuple, Optional, Union
 
 import numpy as np
 import pandas as pd
@@ -16,11 +16,11 @@ from pam.planner.od import OD
 from pam.planner.utils_planner import (
     apply_mode_to_home_chain,
     calculate_mnl_probabilities,
-    sample_weighted,
-    get_first_leg_time_ratio,
-    get_act_names,
     convert_single_anchor_roundtrip,
-    get_trip_chains_either_anchor
+    get_act_names,
+    get_first_leg_time_ratio,
+    get_trip_chains_either_anchor,
+    sample_weighted,
 )
 from pam.planner.zones import Zones
 
@@ -67,8 +67,7 @@ class SelectionSet:
 
     def sample(self) -> list:
         """Sample from a set of alternative options."""
-        sampled = np.apply_along_axis(
-            func1d=self.func_sampling, axis=1, arr=self.probabilities)
+        sampled = np.apply_along_axis(func1d=self.func_sampling, axis=1, arr=self.probabilities)
         sampled_labels = [self.choice_set.choice_labels[x] for x in sampled]
         self._selections = sampled_labels
         return sampled_labels
@@ -129,8 +128,7 @@ class ChoiceModel:
         link_population(self.population)
         self.od = od
         self.zones = self.parse_zone_data(zones)
-        self.zones.data = self.zones.data.loc[list(
-            od.labels.destination_zones)]
+        self.zones.data = self.zones.data.loc[list(od.labels.destination_zones)]
         self.configuration = ChoiceConfiguration()
         self._selections = None
 
@@ -214,8 +212,7 @@ class ChoiceModel:
 
         idxs = []
         u_choices = []
-        choice_labels = list(itertools.product(
-            od.labels.destination_zones, od.labels.mode))
+        choice_labels = list(itertools.product(od.labels.destination_zones, od.labels.mode))
         choice_labels = [ChoiceLabel(*x) for x in choice_labels]
 
         # iterate across activities
@@ -284,15 +281,9 @@ class DiscretionaryTrips:
             act_names = get_act_names(trip_chain)
             if len(act_names) > 2:
                 if act_names[0] != act_names[-1]:
-                    DiscretionaryTripOD(
-                        trip_chain=trip_chain,
-                        od=self._od
-                    ).update_plan()
+                    DiscretionaryTripOD(trip_chain=trip_chain, od=self._od).update_plan()
                 else:
-                    DiscretionaryTripRound(
-                        trip_chain=trip_chain,
-                        od=self._od
-                    ).update_plan()
+                    DiscretionaryTripRound(trip_chain=trip_chain, od=self._od).update_plan()
 
 
 class DiscretionaryTrip:
@@ -306,16 +297,12 @@ class DiscretionaryTrip:
         c) O->discretionary->discretionary->O
         d) O->discretionary->discretionary->D
 
-    Chains with multiple discretionary trips are solved 
+    Chains with multiple discretionary trips are solved
         recursively, updating the first location each time,
-        and then keeping it fixed as we solve downstream.       
+        and then keeping it fixed as we solve downstream.
     """
 
-    def __init__(
-        self,
-        trip_chain: List[Union[Activity, Leg]],
-        od: OD
-    ) -> None:
+    def __init__(self, trip_chain: List[Union[Activity, Leg]], od: OD) -> None:
         self._trip_chain = trip_chain
         self._od = od
         self.act_names = get_act_names(trip_chain)
@@ -335,7 +322,7 @@ class DiscretionaryTrip:
 
[…]
