+++ tests/conftest.py
@@ -1160,12 +1160,48 @@ def od_discretionary():
     zone_labels = ("h", "b", "w")
     od = ODFactory.from_matrices(
         [
-            ODMatrix("time", "car", zone_labels, zone_labels, np.array([[20, 30, 40], [30, 10, 30], [40, 30, 20]])),
-            ODMatrix("time", "bus", zone_labels, zone_labels, np.array([[30, 40, 45], [40, 10, 40], [45, 40, 30]])),
-            ODMatrix("distance", "car", zone_labels, zone_labels, np.array([[5, 6, 8], [6, 2, 6], [8, 6, 5]])),
-            ODMatrix("distance", "bus", zone_labels, zone_labels, np.array([[5, 7, 9], [7, 2, 7], [9, 7, 5]])),
-            ODMatrix("od_probs", "car", zone_labels, zone_labels, np.array([[500, 333, 250], [333, 1000, 333], [250, 333, 500]])),
-            ODMatrix("od_probs", "bus", zone_labels, zone_labels, np.array([[333, 250, 222], [250, 1000, 250], [222, 250, 333]])),
+            ODMatrix(
+                "time",
+                "car",
+                zone_labels,
+                zone_labels,
+                np.array([[20, 30, 40], [30, 10, 30], [40, 30, 20]]),
+            ),
+            ODMatrix(
+                "time",
+                "bus",
+                zone_labels,
+                zone_labels,
+                np.array([[30, 40, 45], [40, 10, 40], [45, 40, 30]]),
+            ),
+            ODMatrix(
+                "distance",
+                "car",
+                zone_labels,
+                zone_labels,
+                np.array([[5, 6, 8], [6, 2, 6], [8, 6, 5]]),
+            ),
+            ODMatrix(
+                "distance",
+                "bus",
+                zone_labels,
+                zone_labels,
+                np.array([[5, 7, 9], [7, 2, 7], [9, 7, 5]]),
+            ),
+            ODMatrix(
+                "od_probs",
+                "car",
+                zone_labels,
+                zone_labels,
+                np.array([[500, 333, 250], [333, 1000, 333], [250, 333, 500]]),
+            ),
+            ODMatrix(
+                "od_probs",
+                "bus",
+                zone_labels,
+                zone_labels,
+                np.array([[333, 250, 222], [250, 1000, 250], [222, 250, 333]]),
+            ),
         ]
     )
     return od
@@ -1173,411 +1209,833 @@ def od_discretionary():
 
 @pytest.fixture
 def plan_home_work_home():
-    plan = Plan(
-        home_area='h',
-        home_location=Location(loc=Point(0, 0), area='h')
+    plan = Plan(home_area="h", home_location=Location(loc=Point(0, 0), area="h"))
+    plan.add(
+        Activity(
+            seq=1, act="home", area="h", loc=Point(0, 0), start_time=mtdt(0), end_time=mtdt(5 * 60)
+        )
     )
-    plan.add(Activity(
-        seq=1, act='home', area='h', loc=Point(0, 0),
-        start_time=mtdt(0), end_time=mtdt(5 * 60)
-    ))
-    plan.add(Leg(
-        seq=1, mode='car', start_area='h', end_area='w',
-        start_loc=Point(0, 0), end_loc=Point(1, 2),
-        start_time=mtdt(5 * 60), end_time=mtdt(6 * 60)
-    ))
-    plan.add(Activity(
-        seq=2, act='work', area='w', loc=Point(1, 2),
[…]

[…]
