### 2023-09-18 7f82957
assigning speed in pam.variables
