### 2023-09-18 b588354
reverting a logic change in d_zone_sample_choice
