### 2023-09-18 2c428da
Updates following review.
Improve tests, add auto font color detection.
