+++ tests/test_10_plotting.py
@@ -29,6 +29,7 @@ from pam.plot.stats import (
     time_binner,
 )
 from pam.policy import policies
+from pam.variables import DEFAULT_ACTIVITIES_FONTSIZE, DEFAULT_ACTIVITIES_PLOT_WIDTH
 
 
 @pytest.fixture
@@ -195,16 +196,19 @@ def test_plot_activities_no_legend(person_df):
     assert not any(isinstance(i, Legend) for i in fig.get_children())
 
 
-def test_plot_activities_with_cmap(person_df):
+def test_plot_activities_user_defined_cmap(person_df):
     cmap = {"Home": (1, 1, 1), "Education": (0, 0, 0), "Travel": (0.3, 0.3, 0.3)}
+    default_opacity = 1
     fig, ax = plot_activities(person_df, cmap=cmap)
     for idx, patch in enumerate(ax.patches):
         label = ax.texts[idx].get_text()
         fc = patch.get_facecolor()
-        assert cmap[label] + (1,) == fc
+        # assertion is against rgba values (i.e., including an opacity value)
+        assert cmap[label] + (default_opacity,) == fc
 
 
-def test_plot_activities_fontcolor(person_df):
+def test_plot_activities_expected_auto_fontcolor(person_df):
+    "Label fontcolour is selected based on perceived luminance of background colour"
     cmap = {"Home": (1, 1, 1), "Education": (0, 0, 0), "Travel": (0.3, 0.3, 0.3)}
     expected_fontcolor = {"Home": "black", "Education": "white", "Travel": "white"}
     fig, ax = plot_activities(person_df, cmap=cmap)
@@ -212,26 +216,27 @@ def test_plot_activities_fontcolor(person_df):
         assert text.get_color() == expected_fontcolor[text.get_text()]
 
 
-def test_plot_activities_with_label_fontsize_partial(person_df):
+def test_plot_activities_non_default_label_fontsize_partial(person_df):
+    "All undefined fontsizes should default to the system default (scaled if fig width has changed from the default)"
     fontsizes = {"Home": 20}
     fig, ax = plot_activities(person_df, label_fontsize={"Home": 20})
     for text in ax.texts:
-        assert text.get_fontsize() == fontsizes.get(text.get_text(), 10)
+        assert text.get_fontsize() == fontsizes.get(text.get_text(), DEFAULT_ACTIVITIES_FONTSIZE)
 
 
-def test_plot_activities_with_label_fontsize_full(person_df):
+def test_plot_activities_non_default_label_fontsize_all(person_df):
     fontsizes = {"Home": 20, "Education": 5, "Travel": 15}
     fig, ax = plot_activities(person_df, label_fontsize=fontsizes)
     for text in ax.texts:
         assert text.get_fontsize() == fontsizes[text.get_text()]
 
 
-def test_plot_activities_with_new_width(person_df):
+def test_plot_activities_non_default_fig_width(person_df):
     width = 40
-    scaled_fontsize = 10 * width / 16  # default width is 16, so scaled proportional to it
-    fig, ax = plot_activities(person_df, width=40)
+    scaled_fontsize = DEFAULT_ACTIVITIES_FONTSIZE * width / DEFAULT_ACTIVITIES_PLOT_WIDTH
+    fig, ax = plot_activities(person_df, width=width)
 
-    assert fig.get_figwidth() == 40
+    assert fig.get_figwidth() == width
     assert ax.title.get_fontsize() == scaled_fontsize
 
 
