+++ pam/variables.py
@@ -21,3 +21,6 @@ TRANSIT_MODES = ["bus", "rail", "pt", "subway"]  # modes for which the maximum w
 SMALL_VALUE = 0.000001
 
 LONG_TERM_ACTIVITIES = ["work", "school", "education"]
+
+DEFAULT_ACTIVITIES_PLOT_WIDTH = 16
+DEFAULT_ACTIVITIES_FONTSIZE = 10
