### 2023-09-19 9733174
Minor update to `DiscretionaryTrip` ABC
