### 2023-09-19 5506c6a
expose tolerance parameters
