### 2023-09-21 66e8678
adding tests for apply method
