### 2023-09-22 e51de41
Bump to v0.3.0; clean changelog; rerun notebooks
