+++ CHANGELOG.md
@@ -16,14 +16,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.3.0] - 2023-09-22
 
 ### Fixed
 - Fix readme CI badge ([#248])
-- Fix for cropping as per issue [#241](https://github.com/arup-group/pam/issues/241) ([#240]).
-- optimise.grid.grid_search fixed ([#239]).
-- `TourPlanner` prevents sampling of duplicate destinations, and prevents origin being sampled as a destination ([#231]).
-- Fix for [#221](https://github.com/arup-group/pam/issues/221), improved "pt simplification" ([#222])
+- Fix for plan cropping as per issue [#241] ([#240]).
+- [][pam.optimise.grid.grid_search] fixed ([#239]).
+- [`TourPlanner`][pam.samplers.tour.TourPlanner] prevents sampling of duplicate destinations, and prevents origin being sampled as a destination ([#231]).
+- Fix for [][pam.activity.Plan.simplify_pt_trips] as per issue [#221], improving "pt simplification" ([#222])
 
 ### Added
 - MATSim warm starting example ([#239]).
@@ -31,33 +31,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Anaconda package of PAM, available on the `city-modelling-lab` channel ([#211]).
 - Python versions 3.9 to 3.11 support ([#192], [#210]).
 - Documentation, now available at https://arup-group.github.io/pam ([#197]).
-- Time-space prism method for selecting the location of non-mandatory activities ([#252](https://github.com/arup-group/pam/pull/252)).
+- Time-space prism method for selecting the location of non-mandatory activities ([#252]).
 - Simple IPF approach for generating synthetic populations ([#253]).
+- More control over the look of activity plan plots, with keyword arguments for [][pam.plot.plot_activities] (accessed via e.g. [pam.core.Person.plot]) extended to allow figure width and per-activity label font sizes to be updated ([#249]).
 - **internal** [Codecov](https://codecov.io) and [pre-commit](https://pre-commit.ci/) CI bots ([#202]).
-- **internal** Github action job to build PAM and run tests on a Windows machine ([#192]).
+- **internal** Github action job to build PAM and run tests on Windows and MacOS machines ([#192]).
 - **internal** Contribution guidelines and issue/pull request templates ([#207]).
 
 ### Changed
 - Documentation and examples improved ([#239]).
-- `ActivityDuration` class name changed to `DurationEstimator` ([#231]).
-- `TourPlanner` class sequences stops using GreedyTSP algorithm, rather than previous method sorting by distance from depot ([#231]).
+- [`TourPlanner`][pam.samplers.tour.TourPlanner] class sequences stops using GreedyTSP algorithm, rather than previous method sorting by distance from depot ([#231]).
 - Minor changes to docs for zsh users (eg `pip install '.[dev]'`)([#219]).
 - Recommended installation instructions, to use [mamba](https://mamba.readthedocs.io/en/latest/index.html) instead of pip ([#192], [#211]).
+- Docker image entry point from python to bash ([#230]).
 - **internal** Source code and example notebook code layout to align with pep8 guidelines and to remove unused dependency imports ([#196], [#201]).
+- **internal** Docstrings to Google style ([#208])
[…]
