### 2023-09-22 f27f19c
Clean up docstrings; add inherited members to API
