+++ pam/planner/choice_location.py
@@ -261,20 +261,18 @@ class ChoiceMNL(ChoiceModel):
 
 
 class DiscretionaryTrips:
-    """
-    Solve discretionary trip location choice of a PAM plan
-    """
-
     def __init__(self, plan: Plan, od: OD) -> None:
+        """Solve discretionary trip location choice of a PAM plan.
+
+        Args:
+            plan (Plan): PAM plan.
+            od (OD): An object holding origin-destination matrices.
+        """
         self._plan = plan
         self._od = od
 
     def update_plan(self):
-        """Update the locations (in-place) of each non-mandatory activity location in the plan.
-
-        Returns:
-            None
-        """
+        """Update the locations (in-place) of each non-mandatory activity location in the plan."""
         trip_chains = get_trip_chains_either_anchor(self._plan)
         for trip_chain in trip_chains:
             # if only one achor, convert to round-trip
@@ -289,22 +287,20 @@ class DiscretionaryTrips:
 
 class DiscretionaryTrip(ABC):
     def __init__(self, trip_chain: list[Union[Activity, Leg]], od: OD) -> None:
-        """
-        Location choice for discretionary trips
-            in a trip chain
+        """Location choice for discretionary trips in a trip chain.
 
         Cases:
-            a) O->discretionary->O (DiscretionaryTripRound)
-            b) O->discretionary->D (DiscretionaryTripOD)
-            c) O->discretionary->discretionary->O
-            d) O->discretionary->discretionary->D
+            1. O->discretionary->O (DiscretionaryTripRound)
+            2. O->discretionary->D (DiscretionaryTripOD)
+            3. O->discretionary->discretionary->O
+            4. O->discretionary->discretionary->D
 
-        Chains with multiple discretionary trips are solved
-            recursively, updating the first location each time,
-            and then keeping it fixed as we solve downstream.
+        Chains with multiple discretionary trips are solved recursively,
+            updating the first location each time, and then keeping it fixed as we solve downstream.
 
         Args:
-            list[Union[Activity, Leg]]: A trip chain between two long-term activities.
+            trip_chain (list[Union[Activity, Leg]]): A trip chain between two long-term activities.
+            od (OD): An object holding origin-destination matrices.
         """
         self._trip_chain = trip_chain
         self._od = od
@@ -326,8 +322,8 @@ class DiscretionaryTrip(ABC):
             raise TypeError("Each even element in the trip chain should be an activity")
 
     @abstractmethod
-    def choose_destination(self):
-        """Selects a destination for the discretionary activity
+    def choose_destination(self) -> str:
+        """Selects a destination for the discretionary activity.
 
         Returns:
             str: Selected destination zone name.
@@ -335,13 +331,8 @@ class DiscretionaryTrip(ABC):
 
     def update_plan(self):
         """
-        Update the PAM activity locations of the first activity activity
-            in the trip chain,
-            and continue downstream until the entire chain is solved.
-
-        Returns:
-            None
-
+        Update the PAM activity locations of the first activity in the trip chain,
+        and continue downstream until the entire chain is solved.
         """
         if len(self.act_names) > 2:
             # update locations
@@ -362,20 +353,23 @@ class DiscretionaryTrip(ABC):
                 DiscretionaryTripOD(trip_chain=self._trip_chain[2:], od=self._od).update_plan()
 
     @property
-    def od(self):
+    def od(self) -> OD:
         return self._od
 
 
 class DiscretionaryTripRound(DiscretionaryTrip):
     """
-    Location choice for a single discretionary trip,
-        where we have the same anchor at the start and end of the chain.
+    Location choice for a single discretionary trip, where we have the same anchor at the start and end of the chain.
 
[…]
