### 2023-09-22 5cddbe0
Silence a pandas deprecation warning.
