+++ tests/test_100_memory_profiling.py
@@ -6,11 +6,13 @@ import pytest
 from pam import read
 
 BENCHMARK_MEM = "2890.59 MB"
+BENCHMARK_SECONDS = 200
 
 data_dir = Path(__file__).parent / "test_data"
 
 
 @pytest.mark.limit_memory(BENCHMARK_MEM)
+@pytest.mark.timeout(BENCHMARK_SECONDS)
 @pytest.mark.high_mem
 def test_activity_loader():
     trips = pd.read_csv(data_dir / "extended_travel_diaries.csv.gz")
