### 2023-09-25 f45da39
Assign seq when not present
