### 2023-09-25 1830e45
Try with more lenient timeout benchmark.
memray has a time overhead that is difficult to gauge on CI runners
