### 2023-09-26 6e009e1
Split out time and memory profiling
