+++ tests/test_100_memory_profiling.py
@@ -5,17 +5,28 @@ import pytest
 
 from pam import read
 
-BENCHMARK_MEM = "2890.59 MB"
-BENCHMARK_SECONDS = 300
+BENCHMARK_MEM = "1400 MB"
+BENCHMARK_SECONDS = 90
 
 data_dir = Path(__file__).parent / "test_data"
 
+pytest.fixture(scope="module")
 
-@pytest.mark.limit_memory(BENCHMARK_MEM)
-@pytest.mark.timeout(BENCHMARK_SECONDS)
-@pytest.mark.high_mem
-def test_activity_loader():
+
+def trips_attrs():
     trips = pd.read_csv(data_dir / "extended_travel_diaries.csv.gz")
     attributes = pd.read_csv(data_dir / "extended_persons_data.csv.gz")
     attributes.set_index("pid", inplace=True)
-    read.load_travel_diary(trips, attributes)
+    return trips, attributes
+
+
+@pytest.mark.limit_memory(BENCHMARK_MEM)
+@pytest.mark.high_mem
+def test_activity_loader_mem(trips_attrs):
+    read.load_travel_diary(*trips_attrs)
+
+
+@pytest.mark.timeout(BENCHMARK_SECONDS)
+@pytest.mark.high_mem
+def test_activity_loader_time(trips_attrs):
+    read.load_travel_diary(*trips_attrs)
