### 2023-09-26 6d2c670
Fix fixture decorator
