### 2023-09-26 61ef095
Further loosen timing benchmark
Locally it takes ~75 seconds.
On a CI runner it seems to take just over 200 seconds.
To be safe, we go for 250 seconds.
