+++ v0.3.0/reference/pam/write/diary/diary.md
@@ -1,0 +1,1 @@
+::: pam.write.diary
