+++ v0.3.0/reference/pam/policy/probability_samplers/probability_samplers.md
@@ -1,0 +1,1 @@
+::: pam.policy.probability_samplers
