+++ v0.3.0/reference/pam/activity/activity.md
@@ -1,0 +1,1 @@
+::: pam.activity
