+++ v0.3.0/reference/pam/report/benchmarks/benchmarks.md
@@ -1,0 +1,1 @@
+::: pam.report.benchmarks
