+++ v0.3.0/reference/pam/samplers/facility/facility.md
@@ -1,0 +1,1 @@
+::: pam.samplers.facility
