+++ v0.3.0/reference/pam/operations/cropping/cropping.md
@@ -1,0 +1,1 @@
+::: pam.operations.cropping
