+++ v0.3.0/reference/pam/samplers/spatial/spatial.md
@@ -1,0 +1,1 @@
+::: pam.samplers.spatial

[…]
