### 2023-10-02 #5845 @martinfleis
Hi @fredshone 👋, I'll be the editor if this submission. If you'll have any queries, send them my way.

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. Your submission is slightly off my expertise, so any help with finding suitable reviewers would help. Thanks!
