### 2023-10-02 #5845 @martinfleis
@fredshone the paper is a bit on the short side. Could you incorporate these two suggestions before I'll send it to reviewers?

1. It should include the description of the state of the field. You should describe how this software compares to other commonly-used packages.
2. Can you include a clear description of the high-level functionality of the package? You covered the purpose but it is not clear how the package works.

Thank you!
