### 2023-10-02 0773d57
adds sections following initial review
