+++ resources/paper/paper.bib
@@ -50,3 +50,34 @@
     year = 2023,
     url = {https://aetransport.org/past-etc-papers/conference-papers-2022?abstractId=7527&state=b}
 }
+
+@article{ActivitySim,
+  title = {ActivitySim: large-scale agent based activity generation for infrastructure simulation},
+  author = {Gali, Emmanuel and Eidenbenz, Stephan and Mniszewski, Sue and Cuellar, Leticia and Teuscher, Christof},
+  abstractNote = {The United States' Department of Homeland Security aims to model, simulate, and analyze critical infrastructure and their interdependencies across multiple sectors such as electric power, telecommunications, water distribution, transportation, etc. We introduce ActivitySim, an activity simulator for a population of millions of individual agents each characterized by a set of demographic attributes that is based on US census data. ActivitySim generates daily schedules for each agent that consists of a sequence of activities, such as sleeping, shopping, working etc., each being scheduled at a geographic location, such as businesses or private residences that is appropriate for the activity type and for the personal situation of the agent. ActivitySim has been developed as part of a larger effort to understand the interdependencies among national infrastructure networks and their demand profiles that emerge from the different activities of individuals in baseline scenarios as well as emergency scenarios, such as hurricane evacuations. We present the scalable software engineering principles underlying ActivitySim, the socia-technical modeling paradigms that drive the activity generation, and proof-of-principle results for a scenario in the Twin Cities, MN area of 2.6 M agents.},
+  doi = {},
+  url = {https://www.osti.gov/biblio/960770}, journal = {},
+  number = ,
+  volume = ,
+  place = {United States},
+  year = {2008},
+  month = {1}
+} 
+
+@article{Eqasim,
+  title = {Introducing the Eqasim Pipeline: {{From}} Raw Data to Agent-Based Transport Simulation},
+  shorttitle = {Introducing the Eqasim Pipeline},
+  author = {H{\"o}rl, Sebastian and Balac, Milos},
+  year = {2021},
+  month = jan,
+  journal = {Procedia Computer Science},
+  series = {The 12th {{International Conference}} on {{Ambient Systems}}, {{Networks}} and {{Technologies}} ({{ANT}}) / {{The}} 4th {{International Conference}} on {{Emerging Data}} and {{Industry}} 4.0 ({{EDI40}}) / {{Affiliated Workshops}}},
+  volume = {184},
+  pages = {712--719},
+  issn = {1877-0509},
+  doi = {10.1016/j.procs.2021.03.089},
+  urldate = {2023-10-02},
+  abstract = {This paper introduces the eqasim framework with the aim to provide a consistent pipeline from raw data to a final transport simulation. It therefore lays the foundation to achieve fully reproducible agent-based transport simulations. While the pathway from raw data to a generic synthetic travel demand was covered previously for specific use cases, here the general methodology is summarized. Furthermore, the tools and methods for combining MATSim simulations with flexibly definable discrete choice models is described, which is the core of the existing simulation implementations of eqasim for \^ile-de-France, Switzerland, Sao Paulo and California.},
+  keywords = {agent-based,eqasim,open data,open source,reproducibility,simulation,transport},
+  file = {/Users/fredshone/Zotero/storage/K6KJ42V2/Hörl and Balac - 2021 - Introducing the eqasim pipeline From raw data to .pdf;/Users/fredshone/Zotero/storage/796NY3UI/S1877050921007274.html}
+}
