### 2023-10-02 #5845 @martinfleis
@fredshone FYI, only one command for `@editorialbot` is possible per comment.
