+++ resources/paper/paper.bib
@@ -51,18 +51,19 @@
     url = {https://aetransport.org/past-etc-papers/conference-papers-2022?abstractId=7527&state=b}
 }
 
-@article{ActivitySim,
-  title = {ActivitySim: large-scale agent based activity generation for infrastructure simulation},
-  author = {Gali, Emmanuel and Eidenbenz, Stephan and Mniszewski, Sue and Cuellar, Leticia and Teuscher, Christof},
-  abstractNote = {The United States' Department of Homeland Security aims to model, simulate, and analyze critical infrastructure and their interdependencies across multiple sectors such as electric power, telecommunications, water distribution, transportation, etc. We introduce ActivitySim, an activity simulator for a population of millions of individual agents each characterized by a set of demographic attributes that is based on US census data. ActivitySim generates daily schedules for each agent that consists of a sequence of activities, such as sleeping, shopping, working etc., each being scheduled at a geographic location, such as businesses or private residences that is appropriate for the activity type and for the personal situation of the agent. ActivitySim has been developed as part of a larger effort to understand the interdependencies among national infrastructure networks and their demand profiles that emerge from the different activities of individuals in baseline scenarios as well as emergency scenarios, such as hurricane evacuations. We present the scalable software engineering principles underlying ActivitySim, the socia-technical modeling paradigms that drive the activity generation, and proof-of-principle results for a scenario in the Twin Cities, MN area of 2.6 M agents.},
-  doi = {},
-  url = {https://www.osti.gov/biblio/960770}, journal = {},
-  number = ,
-  volume = ,
-  place = {United States},
-  year = {2008},
-  month = {1}
-} 
+@inproceedings{ActivitySim,
+  author = {Galli, Emanuele and Cu\'{e}llar, Leticia and Eidenbenz, Stephan and Ewers, Mary and Mniszewski, Sue and Teuscher, Christof},
+  title = {ActivitySim: Large-Scale Agent-Based Activity Generation for Infrastructure Simulation},
+  year = {2009},
+  publisher = {Society for Computer Simulation International},
+  address = {San Diego, CA, USA},
+  abstract = {We introduce ActivitySim, a simulator that generates daily activities for a population of millions of individual agents, each characterized by a set of demographic attributes that is based on US census data. The daily schedules for each agent that consist of a sequence of typical activities, such as sleeping, shopping, working etc., with their respecting starting and ending times. Further, each of the scheduled activities is geo-located at an appropriate location for the type of activity and for the personal situation of the agent, like a business or private residence. ActivitySim has been developed as part of a larger effort to understand the interdependencies among national infrastructure networks and their demand profiles that emerge from the different activities of individuals in both baseline as well as emergency scenarios, such as hurricane evacuations. We present the scalable software engineering principles underlying ActivitySim, the socio-technical modeling paradigms that drive the activity generation, and proof-of-principle results for a 2.6 M agent scenario in the Twin Cities, MN, area.},
+  booktitle = {Proceedings of the 2009 Spring Simulation Multiconference},
+  articleno = {16},
+  numpages = {9},
+  location = {San Diego, California},
+  series = {SpringSim '09}
+}
 
 @article{Eqasim,
   title = {Introducing the Eqasim Pipeline: {{From}} Raw Data to Agent-Based Transport Simulation},
