+++ resources/paper/paper.bib
@@ -1,5 +1,5 @@
 @book{MATSim:2016,
-  	Author = {{Horni}, A. and {Nagel}, K. and {Axhausen}, K.W. (eds.)},
+  	Author = {{Horni}, A. AND {Nagel}, K. AND {Axhausen}, K.W. (eds.)},
   	Booktitle = {2016 The Multi-Agent Transport Simulation MATSim. License: CC-BY 4.0},
   	Publisher = {London: Ubiquity Press.},
   	Title = {{The Multi-Agent Transport Simulation MATSim}},
@@ -18,7 +18,7 @@
 }
 
 @book{Rasouli,
-  	Author = {Rasouli, S. and Timmermans, H},
+  	Author = {Rasouli, S. AND Timmermans, H},
   	Title = {Activity-based models of travel demand: promises, progress and prospects},
   	Publisher = {International Journal of Urban Sciences},
   	Year = 2014,
@@ -27,7 +27,7 @@
 }
 
 @article{Episim,
-    author = {Müller S. A. and Balmer M. and Neumann A. and Nagel, K.},
+    author = {Müller, S. A. AND Balmer, M. AND Neumann, A. AND Nagel, K.},
     title = "{VSP Working Paper: Mobility traces and spreading of COVID-19}",
     year = 2020,
     doi = {10.14279/depositonce-9835},
@@ -35,7 +35,7 @@
 }
 
 @misc{Medium,
-  author = {Shone F. and Kozlowska K.},
+  author = {Shone, F. AND Kozlowska, K.},
   title = {Pandemic Activity Modifier: Intro},
   year = {2020},
   publisher = {Medium},
@@ -43,7 +43,7 @@
 }
 
 @article{Castro:2023,
-    author = {Castro D. A. and Ford A. and Palacin R. and James, P.},
+    author = {Castro, D. A. AND Ford, A. AND Palacin, R. AND James, P.},
     title = {European Transport Conference Papers 2022: Agent-based Modelling of Low Carbon Future Mobility},
     seminar = {Day 2 (8 Sep 2022), Session 6, AGENT BASED MODELS},
     keywords = {Agent-based model, sustainability, active travel, low-carbon future mobility, simulation},
@@ -52,7 +52,7 @@
 }
 
 @inproceedings{ActivitySim,
-  author = {Galli, Emanuele and Cu\'{e}llar, Leticia and Eidenbenz, Stephan and Ewers, Mary and Mniszewski, Sue and Teuscher, Christof},
+  author = {Galli, Emanuele AND Cu\'{e}llar, Leticia AND Eidenbenz, Stephan AND Ewers, Mary AND Mniszewski, Sue AND Teuscher, Christof},
   title = {ActivitySim: Large-Scale Agent-Based Activity Generation for Infrastructure Simulation},
   year = {2009},
   publisher = {Society for Computer Simulation International},
@@ -68,7 +68,7 @@
 @article{Eqasim,
   title = {Introducing the Eqasim Pipeline: {{From}} Raw Data to Agent-Based Transport Simulation},
   shorttitle = {Introducing the Eqasim Pipeline},
-  author = {H{\"o}rl, Sebastian and Balac, Milos},
+  author = {H{\"o}rl, Sebastian AND Balac, Milos},
   year = {2021},
   month = jan,
   journal = {Procedia Computer Science},
