### 2023-10-09 #5845 @fredshone
so sorry @martinfleis ! I saw the weird formatting and thought it was some new concise JOSS style... I am new to .bib.

Anyway looks correct now. Thanks for your patience.
