### 2023-10-23 #5845 @martinfleis
Hi @martibosch and @rusu24edward, would you be willing to review this JOSS submission? I believe that it is reasonably close to your expertise. Thanks!
