### 2023-11-01 #5845 @fredshone
@martinfleis I've gone through the list of reviewers again looking for (more relaxed) relevance, some ideas, but not great ones:

- There is some application for energy (demand) modelling - eg Gregor Becker
- There is some application for epidemiology/ABM - eg chrisvoncsefalvay
- Civil Engineering (most common application of pam is transport modelling which is ~Civil Engineering) - eg dpzhuX
- General interest in "engineering" - eg msdaldo
- Data engineering (bit of a stretch as pam is pretty simple) - eg singhpraveen2010

Other than martibosch (🙇🏻 ) there seems a lack of obviously related persons - Is it helpful to propose reviewers from "outside"? I know some likely _**unbiased**_ **_third_** parties.
