+++ pam/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.1.dev"
+__version__ = "0.3.1"
 
 
 class PAMValidationError(Exception):
