+++ v0.3.1/reference/pam/array/decode/decode.md
@@ -1,0 +1,1 @@
+::: pam.array.decode
