+++ v0.3.1/reference/pam/read/diary/diary.md
@@ -1,0 +1,1 @@
+::: pam.read.diary
