+++ v0.3.1/reference/pam/array/distance/distance.md
@@ -1,0 +1,1 @@
+::: pam.array.distance
