+++ v0.3.1/reference/pam/plot/plans/plans.md
@@ -1,0 +1,1 @@
+::: pam.plot.plans
