+++ v0.3.1/reference/pam/location/location.md
@@ -1,0 +1,1 @@
+::: pam.location
