+++ v0.3.1/reference/pam/vehicles/vehicles.md
@@ -1,0 +1,1 @@
+::: pam.vehicles
