### 2023-11-28 #6097 @editorialbot
[REVIEW]: PAM: Population Activity Modeller
**Submitting author:** @fredshone (<a href="http://orcid.org/0009-0008-1079-0081">Fred Shone</a>)
**Repository:** https://github.com/arup-group/pam
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.4
