### 2023-12-12 f043f3b
Try with newer gdal version
