### 2023-12-12 9836ec0
Bump minimum version to py3.10
