+++ .github/workflows/pr-ci.yml
@@ -19,7 +19,7 @@ jobs:
         os: [windows-latest, ubuntu-latest, macos-latest]
         py3version: ["10", "12"]
       fail-fast: false
-    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-install-lint-test.yml@main
+    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-install-lint-test.yml@try-python--m-pytest
     with:
       os: ${{ matrix.os }}
       py3version: ${{ matrix.py3version }}
@@ -29,17 +29,17 @@ jobs:
       upload_to_codecov: false
 
   test-coverage:
-    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-install-lint-test.yml@main
+    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-install-lint-test.yml@try-python--m-pytest
     with:
       os: ubuntu-latest
-      py3version: "11"
+      py3version: "12"
       notebook_kernel: pam
       lint: false
       pytest_args: 'tests/'  # ignore example notebooks
       upload_to_codecov: true
 
   memory-profile:
-    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-memory-profile.yml@main
+    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-memory-profile.yml@try-python--m-pytest
     with:
-      py3version: "11"
+      py3version: "12"
       upload_flamegraph: true
