### 2023-12-12 6b23f41
Revert GDAL pinning
