### 2023-12-12 4a0b76e
Limit GDAL to below v3.8
v3.5 isn't possible with py3.12
