### 2023-12-12 df7c0b3
Test GDAL v3.6
