### 2024-01-08 04c7d6d
Update GDAL pinning; limit tests to windows
