### 2024-01-08 10ed795
Try with older CI (no change in faulthandler)
