+++ .github/workflows/pr-ci.yml
@@ -16,10 +16,10 @@ jobs:
   test:
     strategy:
       matrix:
-        os: [windows-latest, ubuntu-latest, macos-latest]
+        os: [windows-latest]
         py3version: ["9", "10", "11"]
       fail-fast: false
-    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-install-lint-test.yml@try-python--m-pytest
+    uses: arup-group/actions-city-modelling-lab/.github/workflows/python-install-lint-test.yml@bf00346490f8547340c83a58942942bc8458f428
     with:
       os: ${{ matrix.os }}
       py3version: ${{ matrix.py3version }}
