+++ src/pam/__init__.py
@@ -1,4 +1,13 @@
+"""Top-level module for pam."""
+
+import pyproj
+
+__author__ = """Fred Shone"""  # triple quotes in case the name has quotes in it.
+__email__ = "fred.shone@arup.com"
 __version__ = "0.3.1"
+
+
+pyproj.network.set_network_enabled(False)
 
 
 class PAMValidationError(Exception):
