+++ .github/workflows/daily-scheduled-ci.yml
@@ -16,7 +16,7 @@ jobs:
     uses: arup-group/actions-city-modelling-lab/.github/workflows/python-install-lint-test.yml@main
     with:
       os: ubuntu-latest
-      py3version: "11"
+      py3version: "12"
       notebook_kernel: pam
       pytest_args: '--no-cov'  # ignore coverage
       cache_mamba_env: false
