+++ pyproject.toml
@@ -18,7 +18,7 @@ filterwarnings = [
 
 [tool.coverage.run]
 branch = true
-source = ["pam/"]
+source = ["src/"]
 
 [tool.coverage.report]
 fail_under = 89
@@ -31,10 +31,12 @@ output = "reports/coverage/coverage.xml"
 
 [tool.black]
 line-length = 100
-skip_magic_trailing_comma = true
+skip-magic-trailing-comma = true
 
 [tool.ruff]
 line-length = 100
+
+[tool.ruff.lint]
 select = ["E", "F", "I", "Q"]
 
 ignore = [
@@ -48,27 +50,22 @@ exclude = [".*", "__pypackages__", "build", "dist", "venv", "reports/"]
 # Allow unused variables when underscore-prefixed.
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
-[tool.ruff.pydocstyle]
-convention = "google"
+pydocstyle = {convention = "google"}
 
-[tool.ruff.mccabe]
 # Unlike Flake8, default to a complexity level of 10.
-max-complexity = 10
+mccabe = {max-complexity = 10}
 
 # Ignore `E402` (import violations) and `F401` (unused imports) in all `__init__.py` files
-[tool.ruff.per-file-ignores]
-"__init__.py" = ["E402", "F401"]
-"*.ipynb" = ["E402"]
+per-file-ignores = {"__init__.py" = ["E402", "F401"]}
+
+flake8-quotes = {docstring-quotes = "double"}
 
 # Ignore `E402` for all notebooks
 [tool.nbqa.addopts]
 ruff = ["--extend-ignore=E402"]
 
-
-[tool.ruff.flake8-quotes]
-docstring-quotes = "double"
-
 [tool.setuptools.packages.find]
+where = ["src"]
 include = ["pam*"]
 
 [tool.setuptools.package-data]
