### 2024-02-23 4190f05
lint and format code post-template alignment
