+++ tests/test_19_combining.py
@@ -1,7 +1,6 @@
 import os
 
 import pytest
-
 from pam.operations import combine
 
 
