+++ tests/test_05_policy_levels.py
@@ -1,7 +1,6 @@
 import random
 
 import pytest
-
 from pam.activity import Activity
 from pam.policy import filters, modifiers, policies, probability_samplers
 from pam.utils import minutes_to_datetime as mtdt
