### 2024-03-25 arup-group/pam#269 @fredshone
integration with scikit-mobility
### What can be improved?

[scikit-mobility](https://github.com/scikit-mobility/scikit-mobility) provides similar plotting ability (probably better) for plotting "trajectories", roughly equivalent to pam routes or more abstractly sequences of trips. Additionally it provides some trajectory methodologies not in pam.

If more or improved trip/plan plotting features are required in future - we should consider utilizing scikit-mobility, either as a dependency or by serializing to scikit-mobility TrajDataFrame or FlowDataFrame.

### Version

v0.3.0
