### 2024-03-25 arup-group/pam#270 @fredshone
Bloated project size
### What can be improved?

Project size is up to 250M. `.git/` represents the vast majority of this (224M).

I took a look at the commits using [this](https://stackoverflow.com/questions/10622179/how-to-find-identify-large-commits-in-git-history) script, and can see some particularly egregious files lurking in history that we don't need...

```
642749c02fc3   16MiB examples/data/example_data/hk_zones.geojson
f262b44596a4   17MiB examples/data/example_data/hk_facilities_epsg_4326.geojson
8f019e221b4c   20MiB examples/data/londinium_facilities_epsg_27700.geojson
515bfe0fbc4b   20MiB examples/data/londinium_facilities_epsg_27700.geojson
```

These are no longer in use. So i think safe to remove from history via [BFG](https://rtyley.github.io/bfg-repo-cleaner/).

However, you will need to then need all contributing users to make a fresh clone to prevent pushing them back into history. I am not sure how feasible this is now. But it will only get harder...

### Version

v0.3.0
