### 2024-03-25 #6097 @fredshone
Thank you for the thorough review @martibosch.

The revisions in the above paper should address your manuscript corrections:

- fixed missing author
- capitalized Python
- removed repetition
- added a beginning and end paragraph to summary

Regarding the project:

1. Performance
- Certainly restricted by the OO design, but we trade for use-ability and broad applicability.
- A millions agents is about 10G. Read time from tabular or xml is about 5 minutes.
- There has been profiling and reworking of very commonly used features (such as io) to speed them up.
- Functionality in the CLI uses streaming to minimize memory requirements. It is possible to built some parallelization on this streaming approach for appropriate tasks.
- The examples are deliberately minimal as we include them (and their data) in our repo and CI pipeline. This ensures we find out when they break. I don't think it is feasible for small open source projects to long-term maintain examples that use external data. Happy to talk anyone's ear off about this 😃 .
- We do periodically muse on performance and our next option would be to be stricter about data types and write some of the core objects in C/Rust. But this is some long way off.

2. Scikit-mobility:

- Thanks for the link. I agree there is some duplicate functionality.
- I have raised [this issue](https://github.com/arup-group/pam/issues/269) and will discuss with team next week.

3. Repo bloat:

- Thanks for bringing this up. In-fact the project is 250M which is unnecessary.
- Normally a download size (217M) would be compressed and significantly smaller than uncompressed rep size.
- In fact we don't get much compression because we have a bloated git history.
- I have detailed in [this issue](https://github.com/arup-group/pam/issues/270) and will push for a clean up next week.
- Pretty sure the bulk of the bloat is some big data files lurking in our history.
- We actually have checks for this in our pipeline and pre-commit so don't know how these got through.

Many thanks!
