### 2024-03-25 #6097 @brynpickering
To add on project bloat: this doesn't extend to the downloadable package (< 1MB) indexed on [PyPI](https://pypi.org/project/cml-pam/#files) and [Anaconda](https://anaconda.org/city-modelling-lab/cml-pam/files).
