### 2024-03-26 arup-group/pam#270 @brynpickering
I would extend this to:

```shell
7ca30ee07cc6 3694558 examples/AK_test.ipynb
9599fc0cadcd 3826931 examples/AK_test.ipynb
9e529d1a79ac 4149741 examples/04_Example-Create-Population-Londinium.ipynb
d52d29205006 4149748 examples/04_Example-Create-Population-Londinium.ipynb
39965de28c9a 6676924 notebooks/Example-Create-Population-Londinium.ipynb
ca60e0ee8975 6679080 notebooks/Example-Create-Population-Londinium.ipynb
2a93286cbb82 6682550 examples/04_Example-Create-Population-Londinium.ipynb
642749c02fc3 17017366 examples/data/example_data/hk_zones.geojson
f262b44596a4 17455220 examples/data/example_data/hk_facilities_epsg_4326.geojson
8f019e221b4c 20524282 examples/data/londinium_facilities_epsg_27700.geojson
515bfe0fbc4b 20546490 examples/data/londinium_facilities_epsg_27700.geojson
```

That's ~105MB.

The `gh-pages` branch is going to continue to bloat the repo. _Ideally_ we could move our documentation to readthedocs.org, but we have had issues in the past with that domain being randomly blocked by the Arup firewall...
