+++ docs/index.md
@@ -1,3 +1,8 @@
+<!--
+By default, the index will be a copy of your repository README preamble.
+You can replace this cross-reference or append/prepend to it by updating this page.
+-->
+
 --8<-- "README.md:docs"
 
 ## Why Activity Plans?
