+++ pyproject.toml
@@ -50,15 +50,19 @@ exclude = [".*", "__pypackages__", "build", "dist", "venv", "reports/"]
 # Allow unused variables when underscore-prefixed.
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
-pydocstyle = {convention = "google"}
+[tool.ruff.lint.pydocstyle]
+convention = "google"
 
+[tool.ruff.lint.mccabe]
 # Unlike Flake8, default to a complexity level of 10.
-mccabe = {max-complexity = 10}
+max-complexity = 10
 
+[tool.ruff.lint.per-file-ignores]
 # Ignore `E402` (import violations) and `F401` (unused imports) in all `__init__.py` files
-per-file-ignores = {"__init__.py" = ["E402", "F401"]}
+"__init__.py" = ["E402", "F401"]
 
-flake8-quotes = {docstring-quotes = "double"}
+[tool.ruff.lint.flake8-quotes]
+docstring-quotes = "double"
 
 # Ignore `E402` for all notebooks
 [tool.nbqa.addopts]
@@ -69,7 +73,8 @@ where = ["src"]
 include = ["pam*"]
 
 [tool.setuptools.package-data]
-pam = ["fixtures/**/*"]
+# "py.typed" is added by default. It allows `mypy` to register the package as having type hints.
+pam = ["py.typed", "fixtures/**/*"]
 
 [build-system]
 requires = ["setuptools"]
