### 2024-04-02 #6097 @martibosch
Hello @fredshone @brynpickering, thank you for your responses. I believe the paper can now move forward.

Just a final note regarding the real-world examples: I understand that is not practical to use large external datasets in CI. Nonetheless, I would recommend mentioning some benchmarks (e.g., the 1M agents) somewhere in the examples, documentation or readme, so that users have some reference expectations. Otherwise, users may open a relatively large dataset and wonder whether it is taking the expected time or the process is stuck.

Martí
