### 2024-04-02 #6097 @fredshone
makes sense. I've added a note in the tabular read example. [PR](https://github.com/arup-group/pam/pull/271)here. It essentially reads; "this is a small example, real datasets will take longer".

Thanks Marti
