### 2024-04-04 #6097 @martinfleis
Thank you @jamesdamillington and @martibosch!

@fredshone please refer to the post-review tasks listed in https://github.com/openjournals/joss-reviews/issues/6097#issuecomment-2036442926 and let me know once you're done with them. Let me know if you have any questions.
