+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Python version 3.12 support
+- Support for `geopandas` v0.15, `gdal` >= v3.5, and `python-Levenshtein` v0.26 ([#268]).
 
 ### Changed
 
@@ -134,6 +136,7 @@ This is the first version of PAM which follows semantic versioning and can be co
 [v0.2.1]: https://github.com/arup-group/pam/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/arup-group/pam/compare/initial_version...v0.2.0
 
+[#268]: https://github.com/arup-group/pam/pull/268
 [#258]: https://github.com/arup-group/pam/pull/258
 [#253]: https://github.com/arup-group/pam/pull/253
 [#252]: https://github.com/arup-group/pam/pull/252
