+++ src/pam/__init__.py
@@ -4,7 +4,7 @@ import pyproj
 
 __author__ = """Fred Shone"""  # triple quotes in case the name has quotes in it.
 __email__ = "fred.shone@arup.com"
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 
 pyproj.network.set_network_enabled(False)
