+++ CHANGELOG.md
@@ -23,6 +23,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Added
+
+### Changed
+
+## [v0.3.2] - 2024-04-04
+
+New release for publishing in Journal of Open Source Software.
+
+### Fixed
 
 ### Added
 
