+++ v0.3.2/static/hooks.py
@@ -1,0 +1,149 @@
+import tempfile
+from pathlib import Path
+
+import mkdocs.plugins
+from mkdocs.structure.files import File
+
+TEMPDIR = tempfile.TemporaryDirectory()
+API_FILES_TO_IGNORE = ["cli.py", "variables.py"]
+
+
+# Bump priority to ensure files are moved before jupyter notebook conversion takes place
+@mkdocs.plugins.event_priority(50)
+def on_files(files: list, config: dict, **kwargs):
+    """Link (1) top-level files to mkdocs files and (2) generate the python API documentation."""
+    for file in sorted(Path("./examples").glob("*.ipynb")):
+        files.append(_new_file(file, config))
+        _get_nav_list(config["nav"], "Examples").append(file.as_posix())
+    for file in Path("./resources").glob("**/*.*"):
+        files.append(_new_file(file, config))
+    files.append(_new_file(Path("./CHANGELOG.md"), config))
+
+    api_nav = _api_gen(files, config)
+    _update_nav(api_nav, config)
+
+    return files
+
+
+def _new_file(path: Path, config: dict, src_dir: str = ".") -> File:
+    """Link a file out in the wilderness to a filename in the documentation directory hierarchy.
+
+    Args:
+        path (Path):
+            Path (relative to "src_dir") to file that you want to bring into the documentation directory.
+            In the documentation directory, this same path with apply (e.g., `[src_dir]/path/to/file.md` will become `[docs_dir]/path/to/file.md`)
+        config (dict): mkdocs config dictionary.
+        src_dir (str, optional):
+            Path in which to find the file you want to bring into the docs directory. Defaults to ".".
+
+    Returns:
+        File: mkdocs object that links your file to the docs directory, ready to be added to the mkdocs file list.
+    """
+    return File(
+        path=path,
+        src_dir=src_dir,
+        dest_dir=config["site_dir"],
+        use_directory_urls=config["use_directory_urls"],
+    )
+
+
+def _api_gen(files: list, config: dict) -> dict:
+    """Project Python API generator
+
+    Args:
+        files (list): mkdocs file list.
+        config (dict): mkdocs config dictionary.
+
+    Returns:
+        dict: Python API navigation tree, to add into the mkdocs `nav` tree.
+    """
+    source_dir = Path(config["watch"][0])
+    source_file = source_dir.parts[-1]
+    api_nav: dict = {"top_level": []}
+    for filepath in sorted(source_dir.rglob("[!_]*.py")):
+        rel_filepath = filepath.relative_to(source_dir)
+        if rel_filepath.as_posix() in API_FILES_TO_IGNORE:
+            continue
+        file = _py_to_md(source_file / rel_filepath, api_nav, config)
+        files.append(file)
+    return api_nav
+
+
+def _py_to_md(filepath: Path, api_nav: dict, config: dict) -> File:
+    """Create a markdown file for the API documentation for a given python file in the package source directory.
+
+    Markdown files are stored in a temporary directory, which will be cleaned after mkdocs has finished building the docs.
+
+    Args:
+        filepath (Path): Path to python file relative to the package source code directory.
+        api_nav (dict): Nested dictionary to fill with mkdocs navigation entries.
+        config (Config): mkdocs config dictionary.
+    Returns:
+        File: mkdocs object that links the temp file to the docs directory, ready to be added to the mkdocs file list.
+    """
+    module_parts = filepath.with_suffix("").parts
+
+    module_name = ".".join(module_parts)
+
+    api_file = "reference" / filepath.with_suffix(".md")
[…]
