+++ v0.3.2/reference/pam/samplers/tour/tour.md
@@ -1,0 +1,1 @@
+::: pam.samplers.tour
