+++ v0.3.2/reference/pam/core/core.md
@@ -1,0 +1,1 @@
+::: pam.core
