+++ v0.3.2/reference/pam/write/matrices/matrices.md
@@ -1,0 +1,1 @@
+::: pam.write.matrices
