+++ v0.3.2/reference/pam/samplers/attributes/attributes.md
@@ -1,0 +1,1 @@
+::: pam.samplers.attributes
