### 2024-04-09 #6097 @fredshone
@martinfleis  - slight delay for a fresh release - I have uploaded to Zenodo:

VERSION: v0.3.2
DOI: 10.5281/zenodo.10948231
URL: https://zenodo.org/doi/10.5281/zenodo.10948231

Title and authors matching paper. 🤞🏻
