### 2024-04-11 #6097 @martinfleis
@fredshone Can you ensure that the list of authors of the Zenodo archive matches the list in the paper? As in proper authors alongside you not using custom fields Zenodo offers.
