### 2024-04-15 #6097 @oliviaguest
@fredshone can you change the reference to be without the double brackets here please:
```
(e.g., (Castro et al., 2023)).
```
