### 2024-04-19 6d87fc3
removes bad ref brackets
