### 2024-04-23 #6097 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
