### 2024-04-23 8faa38b
adds JOSS citation
