### 2024-04-25 #6097 @oliviaguest
Huge thanks to the reviewers and editor, @martinfleis, @jamesdamillington, @martibosch! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors @fredshone! 🥳 🍾
